Move the `quill-util` module under the `@bayou` namespace.
